--- index.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-<<<<<<< HEAD
-    <!-- <meta> -->
-    <!-- <title></title> -->
-
-    <link rel="stylesheet" href="css/base.css">
-
-  </head>
-
-  <body>
-    <p>test</p>
-    <div id="map"></div>
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
-    <script async defer src="https://maps.googleapis.com/maps/api/js?key=AIzaSyA9Pjx18wRYcL0GqoAjlM7H4txb88Qr-n8&libraries=places&callback=googleMap.initMap"></script>
-    <script src="/scripts/app.js"></script>
-    <script src="/scripts/yelp.js"></script>
-=======
-    <title>Food Finder</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="/css/reset.css">
-    <link rel="stylesheet" href="/css/base.css">
-    <link rel="stylesheet" href="/css/layout.css">
-    <link rel="stylesheet" href="/css/modules.css">
-
-    <!-- <-- handlebars script tag here--> -->
-  </head>
-
-  <body>
-
-    <!-- Note: the script files aren't currently in order. 10:13am 2/22 -->
-
-     <script src = "scripts/routes.js"></script>
-     <script src = "vendor/scripts/page.js"></script>
-     <script src = "vendor/scripts/webdb.js"></script>
-     <script src = "vendor/scripts/html5sql.js"></script>
-
-    <!-- CDNs -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.5/handlebars.min.js"></script>
->>>>>>> master
-  </body>
-</html>
