--- scripts/routes.js
@@ -1,8 +1,0 @@
-page('/',
-  articlesController.loadAll,
-  articlesController.index);
-
-page('/about', aboutController.index);
-/// routes will continue once we have information to manipulate. 10:22a 2-22
-
-page();
